app-crypt/certbot-dns-duckdns: update dnspython dependency
Change dnspython dependency from ~=2.1.0 to ~=2.1 to be
exact, effectively depending on major version 2 with minor
version 1 or newer, instead of pinning it on version 2.1.x.

Gentoo has removed =dev-python/dnspython-2.1.0 while
app-crypt/certbot-dns-duckdns depended on that specific version.
By changing the dependency to ~=2.1 this is fixed with regard
to dependency version management, but please open an issue if
you run into issues with the plugin when running Certbot.

Package-Manager: Portage-3.0.30, Repoman-3.0.3
Signed-off-by: Osiris Inferi <gentoo@flut.nl.eu.org>